# storage/engine: reuse iterators for the lifetime of rocksDBBatch