`rocksDBBatch` caches a prefix and normal (non-prefix) iterator upon
first use and reuses them for the lifetime of the batch. This avoids a
cgo call to create an iterator on every request. Note that this change
only affects benchmarks which are doing random inserts and thus not able
to take advantage of the "blind" put optimization.

name                            old time/op    new time/op    delta
TrackChoices1_Cockroach-32         651µs ± 4%     634µs ± 3%  -2.59%  (p=0.000 n=20+20)
TrackChoices10_Cockroach-32        203µs ± 2%     197µs ± 3%  -2.89%  (p=0.000 n=19+20)
TrackChoices100_Cockroach-32       143µs ± 4%     138µs ± 3%  -4.09%  (p=0.000 n=20+20)
TrackChoices1000_Cockroach-32      139µs ± 5%     134µs ± 4%  -3.40%  (p=0.000 n=20+20)
InsertDistinct1_Cockroach-32      2.49ms ± 3%    2.45ms ± 5%  -1.44%  (p=0.028 n=19+20)
InsertDistinct10_Cockroach-32     2.04ms ± 3%    1.96ms ± 2%  -4.03%  (p=0.000 n=20+20)
InsertDistinct100_Cockroach-32    2.41ms ± 3%    2.35ms ± 4%  -2.17%  (p=0.000 n=19+20)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6684)
<!-- Reviewable:end -->