@bdarnell 2016-05-16
Reviewed 2 of 2 files at r1.
Review status: all files reviewed at latest revision, 2 unresolved discussions, all commit checks successful.

---

*[storage/engine/rocksdb.go, line 521 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-KHrYryjEUDTgGpuU-s1:-KHrYrykfnJvSu4XJVia:30524065) ([raw file](https://github.com/cockroachdb/cockroach/blob/dfd13a69e99e637f0747d3f0bef0dd87fd8665d2/storage/engine/rocksdb.go#L521)):*
> ```Go
> 
> func (r *rocksDBBatchIterator) Close() {
> 	// rocksDBBatchIterator.Close() is a no-op. The iterator is left open until
> ```

I'd be more comfortable if we added a boolean `open` flag to `rocksDBBatchIterator` (to be set in NewIterator and cleared in Close) to guard against any attempt to reuse an iterator while another is open.

As the MVCCDeleteRange change above shows, sometimes we have more than one iterator, and it seems like an arbitrary restriction to allow one prefix and one non-prefix. 

---

*[storage/engine/rocksdb.go, line 546 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-KHr_AzzDiBZwMlhYZda:-KHr_AzzDiBZwMlhYZdb:1238269991) ([raw file](https://github.com/cockroachdb/cockroach/blob/dfd13a69e99e637f0747d3f0bef0dd87fd8665d2/storage/engine/rocksdb.go#L546)):*
> ```Go
> func (r *rocksDBBatch) Close() {
> 	if i := &r.prefixIter.rocksDBIterator; i.iter != nil {
> 		C.DBIterDestroy(i.iter)
> ```

Comment that we are deliberately calling C.DBIterDestroy directly instead of going through prefixIter.rocksDBIterator.Close to avoid putting our iterator in the pool.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684)*
<!-- Sent from Reviewable.io -->