@petermattis 2016-05-16
Review status: 1 of 3 files reviewed at latest revision, 2 unresolved discussions, some commit checks pending.

---

*[storage/engine/rocksdb.go, line 521 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-KHrYryjEUDTgGpuU-s1:-KHtXZJZB5DlXBhPsUKq:1257336453) ([raw file](https://github.com/cockroachdb/cockroach/blob/dfd13a69e99e637f0747d3f0bef0dd87fd8665d2/storage/engine/rocksdb.go#L521)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary>
> I'd be more comfortable if we added a boolean `open` flag to `rocksDBBatchIterator` (to be set in NewIterator and cleared in Close) to guard against any attempt to reuse an iterator while another is open.
> 
> As the MVCCDeleteRange change above shows, sometimes we have more than one iterator, and it seems like an arbitrary restriction to allow one prefix and one non-prefix. </details>
Done. I had originally done this then encountered a test error that I didn't want to fix. Fixing the test wasn't as difficult as I thought.

---

*[storage/engine/rocksdb.go, line 546 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-KHr_AzzDiBZwMlhYZda:-KHtX_G2F4Pd_qjinfJN:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/dfd13a69e99e637f0747d3f0bef0dd87fd8665d2/storage/engine/rocksdb.go#L546)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary>
> Comment that we are deliberately calling C.DBIterDestroy directly instead of going through prefixIter.rocksDBIterator.Close to avoid putting our iterator in the pool.</details>
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684)*
<!-- Sent from Reviewable.io -->