@bdarnell 2016-05-16
LGTM. The doc comments for NewIterator should probably say something about this.

<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary>
> #### storage/engine: reuse iterators for the lifetime of rocksDBBatch
> 
> `rocksDBBatch` caches a prefix and normal (non-prefix) iterator upon
> first use and reuses them for the lifetime of the batch. This avoids a
> cgo call to create an iterator on every request. Note that this change
> only affects benchmarks which are doing random inserts and thus not able
> to take advantage of the "blind" put optimization.
> 
> name                            old time/op    new time/op    delta
> TrackChoices1_Cockroach-32         651µs ± 4%     634µs ± 3%  -2.59%  (p=0.000 n=20+20)
> TrackChoices10_Cockroach-32        203µs ± 2%     197µs ± 3%  -2.89%  (p=0.000 n=19+20)
> TrackChoices100_Cockroach-32       143µs ± 4%     138µs ± 3%  -4.09%  (p=0.000 n=20+20)
> TrackChoices1000_Cockroach-32      139µs ± 5%     134µs ± 4%  -3.40%  (p=0.000 n=20+20)
> InsertDistinct1_Cockroach-32      2.49ms ± 3%    2.45ms ± 5%  -1.44%  (p=0.028 n=19+20)
> InsertDistinct10_Cockroach-32     2.04ms ± 3%    1.96ms ± 2%  -4.03%  (p=0.000 n=20+20)
> InsertDistinct100_Cockroach-32    2.41ms ± 3%    2.35ms ± 4%  -2.17%  (p=0.000 n=19+20)
> 
> 
> </details>


---

Review status: 1 of 3 files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---

*[storage/engine/rocksdb_test.go, line 65 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-KHtfCL6Z8bH46y2wj1B:-KHtfCL6Z8bH46y2wj1C:971343688) ([raw file](https://github.com/cockroachdb/cockroach/blob/dcabe9983ce07a78f82177511671d9bee633d9d5/storage/engine/rocksdb_test.go#L65)):*
> ```Go
> 	}
> 
> 	after := b.NewIterator(true)
> ```

Add a comment that `before` and `after` have different values for the `prefix` parameter because that's the only way they can both exist at once. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-:-KHtgvBXnxy8TqIRGekj:-1133796097)*
<!-- Sent from Reviewable.io -->