@petermattis 2016-05-16
Added a comment on `rocksDBBatch.NewIterator`. I also did a little refactoring to share the `rocksDBIterator` initialization and destruction code.

<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary>
> LGTM. The doc comments for NewIterator should probably say something about this.</details>


---

Review status: 1 of 3 files reviewed at latest revision, 3 unresolved discussions, all commit checks successful.

---

*[storage/engine/rocksdb_test.go, line 65 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-KHtfCL6Z8bH46y2wj1B:-KHtlC-oL2O35TEO9Cic:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/dcabe9983ce07a78f82177511671d9bee633d9d5/storage/engine/rocksdb_test.go#L65)):*
<details><summary><i>Previously, bdarnell (Ben Darnell) wrote…</i></summary>
> Add a comment that `before` and `after` have different values for the `prefix` parameter because that's the only way they can both exist at once. </details>
Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6684#-:-KHtlCjE2_CFt-xV8VuE:-160451663)*
<!-- Sent from Reviewable.io -->