@dt 2016-05-16
Closes #5555 ?